# backend/project/swagger.yml
swagger: "2.0"
info:
  description: This is the swagger file that provides the API to fetch data from Qresp server
  version: "1.0"
  title: Swagger REST Article
consumes:
  - "application/json"
produces:
  - "application/json"
basePath: "/api"
...
paths:
...
  /paper/{id}:
    get:
      operationId: "project.api.paper"
      tags:
        - "Paper details"
      summary: "Fetches details of paper given id"
      description: "Fetches paper details given id"
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: Paper Id
      responses:
        200:
          description: "Successful read paper details "
    put:
      operationId: "project.api.update_paper"
      tags:
        - "Paper details"
      summary: "Updates a record's metadata (record owner or admin only)"
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: Paper Id
        - in: body
          name: paper
          description: Partial or full paper metadata; top-level fields are
            merged into the stored record. Server-owned fields (id,
            owner_email, version bookkeeping) are ignored if present.
          required: true
          schema:
            type: object
      responses:
        200:
          description: Record updated
        400:
          description: Payload failed model validation
        401:
          description: Authentication required
        403:
          description: Only the record owner or an admin can edit
        404:
          description: Paper with this id does not exist
  /workflow/{id}:
    get:
      operationId: "project.api.workflow"
      tags:
        - "Workflow details"
      summary: "Fetches workflow of paper given id"
      description: "Fetches workflow of paper given id"
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: Paper Id
      responses:
        200:
          description: "Successful read workflow details of paper"
...
  /dircont:
    post:
      operationId: "project.api.dircontents"
      tags:
        - "Data source contents"
      summary: Returns a list of files and folders inside the directory
      parameters:
        - in: body
          name: req
          description: Contents to get of link.
          schema:
            type: object
            required:
              - link
              - src
              - service
            properties:
              link:
                type: string
              src:
                type: string
              service:
                type: boolean
      responses:
        200:
          description: Contents returned
        500:
          description: Failure
  /preview/{id}:
    get:
      operationId: "project.api.getPreview"
      tags:
        - "Get metadata for the preview"
      summary: "Get the metadata for preview from the id"
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: Metadata paper ID
      responses:
        200:
          description: Metadata successfully returned
        400:
          description: Incorrect paper id, paper with this id does not exist in the database
        500:
          description: Internal server error
  /preview:
    post:
      operationId: "project.api.generatePreview"
      tags:
        - "Generate metadata link"
      summary: "Add data for previewing the metadata under curation"
      parameters:
        - in: body
          name: paper
          description: The metadata for which the preview is to be generated
          schema:
            type: object
      responses:
        200:
          description: Metadata successfully added and returned
        400:
          description: Incorrect paper object
        500:
          description: Internal server error
  /publish:
    post:
      operationId: "project.api.publish"
      tags:
        - "Generate Publish Link"
      summary: "Send an email to the curator with the link to publish the paper (authenticated users only; the session identity becomes the record owner)"
      parameters:
        - in: body
          name: paper
          description: The paper to be published
          schema:
            type: object
      responses:
        200:
          description: Successfully generated the verify link
        400:
          description: Incorrect paper object
        401:
          description: Authentication is required to publish
        500:
          description: Internal server error
  /paper/{id}/raw:
    get:
      operationId: "project.api.raw_paper"
      tags:
        - "Paper details"
      summary: "Returns the stored record document for editing (owner or admin only)"
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: Paper Id
      responses:
        200:
          description: Stored record document returned
        401:
          description: Authentication required
        403:
          description: Only the record owner or an admin may load this
        404:
          description: Paper with this id does not exist
  /account/papers:
    get:
      operationId: "project.api.account_papers"
      tags:
        - "Account"
      summary: "Lists records owned by the current session user"
      responses:
        200:
          description: Compact list of the user's records
        401:
          description: Authentication required
  /account/drafts:
    get:
      operationId: "project.api.account_drafts"
      tags:
        - "Account"
      summary: "Lists the current session user's curator drafts"
      responses:
        200:
          description: Compact list of the user's drafts (no state payloads)
        401:
          description: Authentication required
    post:
      operationId: "project.api.create_account_draft"
      tags:
        - "Account"
      summary: "Saves a new curator draft (state may be arbitrarily incomplete)"
      parameters:
        - in: body
          name: body
          required: true
          description: Draft payload; state is stored as-is, never
            publish-validated.
          schema:
            type: object
            properties:
              title:
                type: string
              state:
                type: object
      responses:
        200:
          description: Draft created; summary returned
        400:
          description: Invalid draft payload
        401:
          description: Authentication required
  /account/drafts/{id}:
    get:
      operationId: "project.api.account_draft"
      tags:
        - "Account"
      summary: "Returns one of the user's drafts including its full state"
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: Draft Id
      responses:
        200:
          description: Draft returned
        401:
          description: Authentication required
        404:
          description: No such draft for this user
    put:
      operationId: "project.api.update_account_draft"
      tags:
        - "Account"
      summary: "Updates one of the user's drafts (title and/or state)"
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: Draft Id
        - in: body
          name: body
          required: true
          description: Fields to update; state is stored as-is.
          schema:
            type: object
            properties:
              title:
                type: string
              state:
                type: object
      responses:
        200:
          description: Draft updated; summary returned
        400:
          description: Invalid draft payload
        401:
          description: Authentication required
        404:
          description: No such draft for this user
    delete:
      operationId: "project.api.delete_account_draft"
      tags:
        - "Account"
      summary: "Deletes one of the user's drafts"
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: Draft Id
      responses:
        200:
          description: Draft deleted
        401:
          description: Authentication required
        404:
          description: No such draft for this user
  /admin/ownerless-papers:
    get:
      operationId: "project.api.ownerless_papers"
      tags:
        - "Admin"
      summary: "Lists legacy records without a verified owner (admin only)"
      responses:
        200:
          description: Compact list of ownerless records
        401:
          description: Authentication required
        403:
          description: Only an admin may list ownerless records
  /paper/{id}/owner:
    put:
      operationId: "project.api.assign_paper_owner"
      tags:
        - "Admin"
      summary: "Assigns a verified owner to a record (admin only)"
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: Paper Id
        - in: body
          name: body
          required: true
          description: The owner to assign; set force=true to replace an
            existing owner.
          schema:
            type: object
            required:
              - owner_email
            properties:
              owner_email:
                type: string
              force:
                type: boolean
      responses:
        200:
          description: Owner assigned
        400:
          description: Invalid email address
        401:
          description: Authentication required
        403:
          description: Only an admin may assign owners
        404:
          description: Paper with this id does not exist
        409:
          description: Record already has a different owner (use force)
  /paper/{id}/permissions:
    get:
      operationId: "project.api.paper_permissions"
      tags:
        - "Paper permissions"
      summary: "Reports whether the current session may edit the record"
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: Paper Id
      responses:
        200:
          description: Permission decision returned
        404:
          description: Paper with this id does not exist
...